Add a separate generate task to rebuild the vgform auto-generated files
Added a separate "generate" task. This task calls "go generate" in the
vgfomr package to (re)generate the autogenerated files. This was
previously part of the "build" task

The new "generate" task is also called from the "all" task, but is not a
prerequsite of the "build" task as the generates files may not differ
from the copies in the repository.

A future revision to the github workflow used for CI/CD will confirm
that the newly generated files match those already in the repository or
fail the build if the generated files are different.