Commit the version of the go.mod before code generation
There are two issues.

Firstly code generation changes the list of imports - by picking up
inports in the *.vugu anf *.go component files.

Secondly the go.mod currently in the repo is the state of the go.mod
_after_ code generation.

This is incorrect, because code generation changes the go.mod in the cas
eof test-002-click (and very likely other tests). As a result the build
fails becuase the go.mod before code generation does not match the
expected state.

This commit commits the go.mod in the state prior to code generation.
The go mod tidy that follows the go generate in the build scripts will
update the go.mod as necessary.

This apears to be the root cause of the Github Actions build failure.