A vg-template example - possible HTML parser bug
This simple example of the use of a vg-template structure appears to
show a failure in the HTML parser.

The templateselect.vugu file as committed will fail to build with:

github.com/vugu/vugu/examples/vg-template
./templateselect_gen.go:65:53: undefined: j
./templateselect_gen.go:66:31: undefined: j
Error: running "go build -o ./main.wasm github.com/vugu/vugu/examples/vg-template" failed with exit code 1

If however the id="failure" <div> is commented out and the id="works"
<div> uncommented the code works as expected.

So there appears to be some interaction that prevents the "j" being
scooped from the <vg-template> into the <option> tag but not when a <p>
tag is used.

Also pulling the <select> tag inside of the <div id="failure"> block
makes no difference.