Add a new test SyscallJsValueOf that calls syscall/js directly
Added a new test SyscallJsValueOf that is a parallel of the original
test VuguJsValueOf.

The test case is the same, but instead of calling vugu/js.ValueOf we
call syscall/js.ValueOf directly.

This test therefore indicates if syscal/js can be used directly with the
current implementation of vugu/js.

This currently results in a build failure:

$ mage testSingleWasmTest github.com/vugu/vugu/wasm-test-suite/test-025-valueof
...
package github.com/vugu/vugu/wasm-test-suite/test-025-valueof
	imports syscall/js: build constraints exclude all Go files in /home/owen/sdk/go1.23.5/src/syscall/js
...
Error: running "go test -v github.com/vugu/vugu/wasm-test-suite/test-025-valueof" failed with exit code 1

This build failure is caused because the underlying `go test` command is being executed with native GOOS and GOARCH values and not GOOS=js GOARCH=wasm.

As per the Go docs a `go test` command will first build the module,
before building the test. In the case of vugu we want the module to be
built with GOOS=js and GOARCH=wasm and not the native values.

These native values then fail the built tag in syscall/js of `//go:build
js && wasm` when then causes the whoel build to fail as per above.

The realted issues are Issue #330 and #328.