An attempt to move wasm test test-002-click to a testdriver approach
This following applies only to the wasm test test-002-click ONLY. The
remaining tests have not yet been moved to the proposed new approach.

How to run.

$ mage -v testSingleWasmTest github.com/vugu/vugu/wasm-test-suite/test-002-click

Attempting to run the full wasm-test-suite via

$ mage -v testwasm

Will fail.

The idea is outlined in this comment to Issue 330
https://github.com/vugu/vugu/issues/330#issuecomment-2612839711

The idea splits a wasm-test into two distinct modules.

The wasm part in the ".../test-002-click" module. This module is ONLY
ever built for a GOOS=js GOARCH=wasm target.

The testdriver part in the "../test-002-click/testdriver/" module. This
module is build via a "go test" command for the native target ONLY.
Since this module contains only test code, we avoid building the
".../test-002-click" module for the native platform, where the module will
never be run or tested against.

The use of modules is the only way to ensure this behaviour. Sub
packages or "main_test" packages do not guarantee this behaviour.

The module split is currently inverted i.e. the test driver module which
contains the Go test code is a sub-module contained in a sub directory of the
module we want to test.

This could be inverted in the future, but would be a more extensive
change.

This commit also greatly improves the behaviour of the

modulesUnderDir(dir string) ([]moduleData, error)

function used by the magefile. See the file "modulerhandler.go".

The original implementation made an implicit assumption that the depth
of the directory tree under the `wasm-test-suite` directory was one. In
other words the `wasm-test-dir` contained the tests, each in their own
directory but that the tests did not contain any further sub directories.
The new version removes this restriction. This was necessary because the
test driver is a sub-directory of the test that it relates to.

Any inversion of this (see above) would require the same change.